+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog and release notes
 
-<!-- ## Unreleased -->
+## Unreleased
 
 <!-- Here goes all the unreleased changes descriptions -->
+
+### Features
+
+- **Breaking Change**: expose shim as a package entry point `type-graphql/shim` (and `/node_modules/type-graphql/build/typings/shim.ts`)
 
 ## v2.0.0-beta.3